dividing by functional area
splitting this out by subject area
playing around with some boilerplate
