+++ src/main.rs
@@ -1,7 +1,9 @@
 use crate::cli::*;
+use crate::data::*;
 use clap::Parser;
 
 pub mod cli;
+pub mod data;
 
 fn main() {
     if let Err(e) = run(Args::parse()) {
